requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.2
Django==3.2.6
django-braces==1.14.0
django-environ==0.4.5
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
httplib2==0.19.1
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
packaging==21.0
pbr==5.6.0
Pillow==8.3.1
platformdirs==2.2.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
wrapt==1.12.1